Move HttpSpanCollector to its own module
similar to Kafka and Scribe, Http should be in its own module, allowing
users to pick whichever they want.

This basically moves https://github.com/openzipkin/brave/commit/2a85ceca5a1b50e8cdfb363a3d82026fea65b205 to its own module
