+++ brave-spancollector-http/README.md
@@ -1,0 +1,4 @@
+# brave-spancollector-http #
+
+SpanCollector that is used to submit spans to Zipkins Http endpoint `/spans`.
+
