+++ brave-spancollector-http/src/main/java/com/github/kristofa/brave/http/HttpSpanCollector.java
@@ -1,5 +1,6 @@
-package com.github.kristofa.brave;
+package com.github.kristofa.brave.http;
 
+import com.github.kristofa.brave.*;
 import com.github.kristofa.brave.internal.Nullable;
 import com.google.auto.value.AutoValue;
 import com.twitter.zipkin.gen.Span;
